# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legend Bucks rewards platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees/{id}/balance:
    get:
      operationId: getEmployeeBalance
      tags: [employees]
      summary: Get an employee's current balance calculated from ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSummary"
...
  /rewards:
    get:
      operationId: listRewards
      tags: [rewards]
      summary: List available rewards
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of rewards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reward"
    post:
      operationId: createReward
      tags: [rewards]
      summary: Create a new reward (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardInput"
      responses:
        "201":
          description: Created reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
        "403":
          description: Forbidden
  /rewards/{id}:
    get:
      operationId: getReward
      tags: [rewards]
      summary: Get a reward by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
        "404":
          description: Not found
    patch:
      operationId: updateReward
      tags: [rewards]
      summary: Update a reward (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardUpdate"
      responses:
        "200":
          description: Updated reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
        "404":
          description: Not found
    delete:
      operationId: deleteReward
      tags: [rewards]
      summary: Delete a reward (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Redemptions
  /redemptions:
    get:
      operationId: listRedemptions
      tags: [redemptions]
      summary: List redemptions
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: pendingApproval
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Redemption"
    post:
      operationId: createRedemption
      tags: [redemptions]
      summary: Redeem a reward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedemptionInput"
      responses:
        "201":
          description: Redemption created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
        "400":
          description: Insufficient balance or out of stock
  /redemptions/{id}:
    get:
      operationId: getRedemption
      tags: [redemptions]
      summary: Get a redemption by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Redemption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
        "404":
          description: Not found
  /redemptions/{id}/approve:
    patch:
      operationId: approveRedemption
      tags: [redemptions]
      summary: Approve a pending redemption (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved redemption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
  /redemptions/{id}/reject:
    patch:
      operationId: rejectRedemption
      tags: [redemptions]
      summary: Reject a redemption and refund bucks (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedemptionDecision"
      responses:
        "200":
          description: Rejected redemption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
  /redemptions/{id}/cancel:
    patch:
      operationId: cancelRedemption
      tags: [redemptions]
      summary: Cancel a redemption and refund bucks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled redemption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
  /redemptions/{id}/fulfill:
    patch:
      operationId: fulfillRedemption
      tags: [redemptions]
      summary: Mark a redemption as fulfilled (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fulfilled redemption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"

  # Team Goals
...
components:
  schemas:
...
    BalanceSummary:
      type: object
      required: [employeeId, balance, totalReceived, totalSpent]
      properties:
        employeeId:
          type: integer
        balance:
          type: integer
        totalReceived:
          type: integer
        totalSpent:
          type: integer
...
    Reward:
      type: object
      required: [id, name, buckCost, active, approvalRequired, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        buckCost:
          type: integer
        imageUrl:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
          description: null means unlimited
        locationRestriction:
          type: ["string", "null"]
        active:
          type: boolean
        approvalRequired:
          type: boolean
        createdAt:
          type: string
    RewardInput:
      type: object
      required: [name, buckCost]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        buckCost:
          type: integer
          minimum: 1
        imageUrl:
          type: string
        quantity:
          type: ["integer", "null"]
        locationRestriction:
          type: ["string", "null"]
        active:
          type: boolean
        approvalRequired:
          type: boolean
    RewardUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        buckCost:
          type: integer
          minimum: 1
        imageUrl:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        locationRestriction:
          type: ["string", "null"]
        active:
          type: boolean
        approvalRequired:
          type: boolean
    Redemption:
      type: object
      required: [id, employeeId, rewardId, status, buckCost, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        rewardId:
          type: integer
        rewardName:
          type: string
        status:
          type: string
          enum: [requested, approved, rejected, fulfilled, cancelled]
        buckCost:
          type: integer
        note:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    RedemptionInput:
      type: object
      required: [rewardId]
      properties:
        rewardId:
          type: integer
        note:
          type: string
    RedemptionDecision:
      type: object
      properties:
        adminNote:
          type: string
...