# lib/api-zod/src/generated/types/redemption.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legend Bucks rewards platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RedemptionStatus } from './redemptionStatus';

export interface Redemption {
  id: number;
  employeeId: number;
  employeeName?: string;
  rewardId: number;
  rewardName?: string;
  status: RedemptionStatus;
  buckCost: number;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt?: string;
}